+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# next
+
+* Added `genericF` for formatting arbitrary data.
+
 # 0.0.0.4
 
 * Added `format` from `text-format`, because in some cases it's nicer than brackets.